Update input.js to have access to ref
In the input's current implementation you lose access to the ref of the input on the form.  This prevents you from using native props like  onSubmitEditing to tab to the next field (ref={n=>this.someField=n} onSubmitEditing={() => this.someField.focus()}) or setting up a custom keyboard and using it to change values of a specific input etc.  Note this is only available in react 16.3+.